crypto_address/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.7.3
bitcoinlib==0.6.10
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cytoolz==0.12.1
Django==4.2.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
fastecdsa==2.3.0
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
lru-dict==1.2.0
MarkupSafe==2.1.2
mnemonic==0.20
multidict==6.0.4
numpy==1.24.3
openapi-codec==1.3.2
packaging==23.1
parsimonious==0.9.0
protobuf==4.23.2
pycryptodome==3.18.0
pyrsistent==0.19.3
pytz==2023.3
regex==2023.5.5
requests==2.31.0
rlp==3.0.0
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
SQLAlchemy==2.0.15
sqlparse==0.4.4
toolz==0.12.0
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==2.0.2
web3==6.4.0
websockets==11.0.3
yarl==1.9.2
